refactored lexer
