--- src/lexer.js
@@ -1,7 +1,0 @@
-'use strict'
-const { Lexer } = require('chevrotain')
-const { lexerDefinition } = require('./tokens')
-
-const ApexLexer = new Lexer(lexerDefinition, { ensureOptimizations: true })
-
-module.exports = ApexLexer
