+++ src/lexer/_shared/helpers.js
