+++ src/lexer/_shared/index.js
