+++ src/lexer/_shared/symbols.js
