+++ src/lexer/apex/classAndMethods.js
