+++ src/lexer/apex/comments.js
