+++ src/lexer/apex/symbols.js
