+++ src/lexer/soql/index.js
@@ -1,0 +1,11 @@
+const { Identifier, WhiteSpace } = require('../shared')
+
+const soqlBasic = require('./soqlBasic')
+const { sharedSymbols } = require('../_shared')
+
+module.exports = {
+  WhiteSpace,
+  ...soqlBasic,
+  Identifier,
+  ...sharedSymbols,
+}
